Add dns_zone var to the elb-dns-registrator job
As part of the DNS migration we need to be able to configure the DNS
zone provided to the Konstructor API. With this change upgrade the
coco/coco-elb-dns-registrator version to 5.1.0. This version support the
DomainZone parameter. Also we add the DomainZone value to be taken from
the global-config configmap dns_zone key.